+++ app/src/main/java/com/djacoronel/gwacalculator/Model/Course.kt
@@ -1,6 +1,6 @@
 package com.djacoronel.gwacalculator.Model
 
 
-class Course(val courseCode: String, val units: Double, var grade: Double) {
+class Course(val courseCode: String, val units: Int, var grade: Double) {
 
 }
